Refactor color management
