+++ flutter/lib/ui/history/history_list_item.dart
@@ -1,8 +1,8 @@
 import 'package:flutter/material.dart';
 
-import 'package:mlperfbench/app_constants.dart';
 import 'package:mlperfbench/data/results/benchmark_result.dart';
 import 'package:mlperfbench/ui/formatter.dart';
+import 'package:mlperfbench/ui/app_styles.dart';
 import 'package:mlperfbench/ui/history/list_item.dart';
 import 'package:mlperfbench/ui/icons.dart';
