+++ flutter/lib/ui/history/utils.dart
@@ -3,11 +3,11 @@ import 'package:flutter/services.dart';
 
 import 'package:bot_toast/bot_toast.dart';
 
-import 'package:mlperfbench/app_constants.dart';
 import 'package:mlperfbench/benchmark/state.dart';
 import 'package:mlperfbench/data/environment/env_android.dart';
 import 'package:mlperfbench/data/environment/environment_info.dart';
 import 'package:mlperfbench/localizations/app_localizations.dart';
+import 'package:mlperfbench/ui/app_styles.dart';
 
 class HistoryHelperUtils {
   final AppLocalizations l10n;
