+++ flutter/lib/ui/settings/snack_bar.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 
-import 'package:mlperfbench/app_constants.dart';
+import 'package:mlperfbench/ui/app_styles.dart';
 
 SnackBar getSnackBar(String message) => SnackBar(
       duration: const Duration(seconds: 2),
